+++ sktime/classification/deep_learning/transformer.py
@@ -6,6 +6,7 @@ from sktime.classification.deep_learning._pytorch import BaseDeepClassifierPytor
 from sktime.utils.dependencies import _check_soft_dependencies
 
 if _check_soft_dependencies("torch", severity="none"):
+    import torch
     from torch.utils.data import DataLoader, Dataset
 else:
 
@@ -14,7 +15,7 @@ else:
 
 
 class TransformerClassifier(BaseDeepClassifierPytorch):
-    """Transformer Classifier based on Pytorch."""
+    """TODO: Fill this."""
 
     _tags = {
         "authors": ["geetu040"],
@@ -96,7 +97,7 @@ class TransformerClassifier(BaseDeepClassifierPytorch):
             }
 
     def _build_network(self, X, y):
-        from sktime.classification.deep_learning.transformer.network import (
+        from sktime.networks.transformer_classifier import (
             TSTransformerEncoderClassiregressor,
         )
 
@@ -121,10 +122,6 @@ class TransformerClassifier(BaseDeepClassifierPytorch):
         )
 
     def _build_dataloader(self, X, y=None):
-        from sktime.classification.deep_learning.transformer.dataset import (
-            PytorchDataset,
-        )
-
         dataset = PytorchDataset(X, y)
         return DataLoader(dataset, self.batch_size)
 
@@ -174,3 +171,39 @@ class TransformerClassifier(BaseDeepClassifierPytorch):
             },
         ]
         return params
+
+
+class PytorchDataset(Dataset):
+    """Dataset specifc to TransformerClassifier."""
+
+    def __init__(self, X, y):
+        # X.shape = (batch_size, n_dims, n_timestamps)
+        X = np.transpose(X, (0, 2, 1))
+        # X.shape = (batch_size, n_timestamps, n_dims)
+
+        self.X = X
+        self.y = y
+
+    def __len__(self):
+        """Get length of dataset."""
+        return len(self.X)
+
+    def __getitem__(self, i):
+        """Get item at index."""
+        x = self.X[i]
+        x = torch.tensor(x, dtype=torch.float)
+        padding_masks = torch.ones(x.shape[:-1], dtype=torch.bool)
+
+        inputs = {
+            "X": x,
+            "padding_masks": padding_masks,
+        }
+
+        # to make it reusable for predict
+        if self.y is None:
+            return inputs
+
+        # return y during fit
+        y = self.y[i]
+        y = torch.tensor(y, dtype=torch.long)
+        return inputs, y
