--- sktime/classification/deep_learning/transformer/__init__.py
@@ -1,3 +1,0 @@
-"""Transformer Classifier."""
-
-from sktime.classification.deep_learning.transformer.model import TransformerClassifier
