+++ sktime/networks/transformer_classifier.py
